Develop to master

+++ backend/src/controllers/personalController.js
@@ -1,5 +1,8 @@
 //here the personal controller
 import { personalModel } from "../models/PersonalModel.js";
+import jsonwebtoken from "jsonwebtoken";
+import { fileURLToPath } from 'url';
+import bcrypt from "bcrypt";
 //READ_ALL - feito
 //READ_ONE - feito
 //CREATE_PERSONAL - feito
@@ -7,6 +10,15 @@ import { personalModel } from "../models/PersonalModel.js";
 //DELETE_PERSONAL - feito
 //LOGIN_PERSONAL - feito
 //LOGOUT_PERSONAL - falta
+// const fs = require('fs');
+// const path = require('path');
+import fs from "fs";
+import path from "path";
+const salt_rounds = process.env.SALT_ROUNDS;
+
+async function crypt(senha){
+    return await bcrypt.hash(senha, Number(salt_rounds));
+}
 
 async function readAll(req, res) {
     try{
@@ -22,7 +34,7 @@ async function readAll(req, res) {
 async function readOne(req, res){
     try{
         //id
-        const personal = await personalModel.findById(req.params.id);
+        const personal = await personalModel.findById(req.params._id);
         if(personal){
             return res.status(200).send(personal);
         }else{
@@ -39,23 +51,9 @@ async function readOne(req, res){
 }
 async function createPersonal(req, res){
     try{
-        //desmembrar objeto
-        const newPersonal = {
-            nome: req.body.nome,
-            CPF: req.body.CPF,
-            email: req.body.email,
-            birth: req.body.birth,
-            senha: req.body.senha,
-            descricao: req.body.descricao,
-            formacao: req.body.formacao,
-            cidade: req.body.cidade,
-            biografia: req.body.biografia,
-            preco: req.body.preco,
-
-        }
         
-
         //verificar email
+        console.log(req.body);
         const existingEmailArray = await personalModel.find({email: req.body.email});
         const existingEmail = existingEmailArray[0];
 
@@ -66,7 +64,7 @@ async function createPersonal(req, res){
         }
 
         //verificar cpf
-        const existingCPFArray = await personalModel.find({CPF: newPersonal.CPF});
+        const existingCPFArray = await personalModel.find({CPF: req.body.CPF});
         const existingCPF = existingCPFArray[0];
 
         if(existingCPF != undefined){
@@ -76,6 +74,27 @@ async function createPersonal(req, res){
         }
 
         //hash de senha
+        const newSenha = await crypt(req.body.senha)
+        // const hashedPassword = await crypt(newPersonal.senha);
+        // newPersonal.senha = hashedPassword;
+        //pegar o arquivo
+        const image = req.file ? req.file.filename : null;;
+        //desmembrar objeto
+        const newPersonal = {
+            nome: req.body.nome, // ok
+            CPF: req.body.CPF, //ok
+            email: req.body.email, //ok
+            birth: req.body.birth, //ok
+            senha: newSenha, //ok
+            descricao: req.body.descricao, // ok
+            formacao: req.body.formacao, //ok
+            cidade: req.body.cidade, //ok
+            biografia: req.body.biografia, //ok
+            preco: req.body.preco,// ok
+            image: image 
+        }
+        
+
         const result = await personalModel.create(newPersonal);
 
         return res.status(201).send({
@@ -85,23 +104,47 @@ async function createPersonal(req, res){
 
     }catch(error){
         return res.status(400).send({
-            error: error,
+            error: error.message,
             message: 'Ocorreu um erro na criação de personal'
         });
     }
 }
 async function updatePersonal(req, res) {
     try{
-        const id = req.params.id;
+        const id = req.params._id;
 
         const update = req.body;
 
-        const updatedPersonal = await personalModel.findByIdAndUpdate(id, update);
+        const updateObject = {
+            CPF: update.CPF,
+            biografia: update.biografia,
+            birth: update.birth,
+            cidade: update.cidade,
+            descricao: update.descricao,
+            email: update.email,
+            formacao: update.formacao,
+            nome: update.nome,
+            preco: update.preco,
+        }
+        if(update.senha.length > 0){
+            const newPassword = await crypt(update.senha);    
+            updateObject.senha = newPassword;
+        }
+        if (req.file){
+            updateObject.image = req.file.filename;
+        }
 
-        return res.status(200).send({
-            message:"Personal atualizado com sucesso.",
-            data: update
-        });
+
+        const updatedPersonal = await personalModel.findByIdAndUpdate(id, updateObject, {returnDocument: "after"});
+
+        console.log(updatedPersonal);
+
+        setTimeout(() => {
+            return res.status(200).send({
+                message:"Personal atualizado com sucesso.",
+                data: updatedPersonal
+            });
+        }, 2000);
 
     }catch(error){
         return res.status(400).send({
@@ -110,23 +153,56 @@ async function updatePersonal(req, res) {
         });
     }
 }
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 async function deletePersonal(req, res){
     try{
-        const id = req.params.id;
+        const id = req.params._id;
+
+        const personal= await personalModel.findById(id);
+        if(!personal){
+            return res.status(400).send({
+                message: "Personal não encontrado"
+            });
+        }
+        const imageName = personal.image;
+
         const personalDelete = await personalModel.findByIdAndDelete(id);
 
         if(personalDelete){
-            return res.status(200).send({
-                message: 'Personal removido com sucesso',
-                data: personalDelete
-            });
+            if(imageName){
+
+                const imagePath = path.resolve(__dirname, '..', '..','..', 'uploads', imageName);
+                console.log(imagePath);
+                fs.unlink(imagePath, (err)=>{
+                if(err){
+                    return res.status(400).send({
+                        message: 'Erro ao excluir imagem do servidor'
+                    });
+                }else{
+    
+                    return res.status(200).send({
+                        message: 'Personal removido com sucesso',
+                        data: personalDelete
+                    });
+                }
+    
+               })
+            }
+            else{
+                return res.status(200).send({
+                    message: 'Personal removido com sucesso',
+                    data: personalDelete
+                });
+            }
+           
+           
         }else{
             return res.status(400).send({
-                message: "Personal não encontrado"
+                message: "Erro ao remover personal"
             })
         }
         
-       
     }catch(error){
         return res.status(400).send({
             message: 'Ocorreu um erro para remover o personal'
@@ -141,6 +217,7 @@ async function loginPersonal(req, res){
 
         const existingPersonalArray = await personalModel.find({email});
         const existingPersonal = existingPersonalArray[0];
+        console.log(existingPersonal);
         if(existingPersonal == undefined){
             return res.status(400).send({
                 message: "Login ou Senha errados.",
@@ -148,20 +225,33 @@ async function loginPersonal(req, res){
             });
         }
         //hash de senha
-        if(senha === existingPersonal.senha){
+        const passwordMatch = await bcrypt.compare(senha, existingPersonal.senha);
+        console.log(passwordMatch);
+        if(!passwordMatch){
+            return res.status(400).send({
+                message: "Login ou Senha errados",
+                status: false
+            })
+        }else{
             console.log(req.body);
+            const token = jsonwebtoken.sign(
+                {
+                    id: existingPersonal._id,
+                    type: "personal" 
+                }, 
+                process.env.SECRET_JWT, 
+                {
+                    expiresIn: '15m'
+                }
+            );
+            const { exp } = jsonwebtoken.decode(token);
             return res.status(200).send({
                 message: "Personal autenticado com sucesso",
                 status: true,
-                token: "personaltoken",
-                personal: existingPersonal
+                token: token,
+                personal: existingPersonal,
+                expiration: exp
             });
-            
-        }else{
-            return res.status(400).send({
-                message: "Login ou Senha errados.",
-                status: false
-            })
         }
     }catch(error){
         return res.status(400).send({
@@ -172,12 +262,23 @@ async function loginPersonal(req, res){
 }
 async function logoutPersonal(req, res){
     try{
+        const token = req.headers.authorization;
 
+        if(token){
+            await blackListModel.create({token});
+        }
+        console.log(token)
+
+        return res.status(200).send({
+            message: "Logout efetuado com sucesso"
+        });
     }catch(error){
-        
+        return res.status(400).send({
+            message: "Ocorreu um erro ao efetuar o logout"
+        })
     }
 }
 
 
 
-export { readAll, readOne, createPersonal, updatePersonal, deletePersonal, loginPersonal };
+export { readAll, readOne, createPersonal, updatePersonal, deletePersonal, loginPersonal, logoutPersonal };
