+++ backend/src/controllers/userController.js
@@ -1,5 +1,17 @@
 //here the user controller
 import { userModel } from "../models/UserModel.js";
+import jsonwebtoken from "jsonwebtoken";
+import bcrypt from "bcrypt";
+import { blackListModel } from "../models/BlackListModel.js";
+import fs from "fs";
+import path from "path";
+import { fileURLToPath } from 'url';
+
+const salt_rounds = process.env.SALT_ROUNDS;
+
+async function crypt(senha){
+    return await bcrypt.hash(senha, Number(salt_rounds));
+}
 //READ_ALL
 //READ_ONE
 //CREATE_USER
@@ -24,6 +36,7 @@ async function readOne(req, res){
         //id
         const user = await userModel.findById(req.params.id);
         if(user){
+
             return res.status(200).send(user);
         }else{
             return res.status(400).send({
@@ -39,37 +52,40 @@ async function readOne(req, res){
 }
 async function createUser(req, res){
     try{
-        //desmembrar objeto
+        //verificar email
+        const existingEmailArray = await userModel.find({email: req.body.email});
+        const existingEmail = existingEmailArray[0];
+        console.log(existingEmailArray);
+        console.log(existingEmail);
+
+        if(existingEmail != undefined || existingEmail != null){
+            // return res.status(400).send({
+            //     message: 'Email ja cadastrado'
+            // });
+            
+            throw new Error("Email já cadastrado");
+        }
+
+        //verificar cpf
+        const existingCPFArray = await userModel.find({CPF: req.body.CPF});
+        const existingCPF = existingCPFArray[0];
+        if(existingCPF){
+            throw new Error("CPF já cadastrado.");
+        }
+
+        const newSenha = await crypt(req.body.senha);
+
+        const image = req.file ? req.file.filename : null;
+
         const newUser = {
             nome: req.body.nome,
             CPF: req.body.CPF,
             email: req.body.email,
             birth: req.body.birth,
-            senha: req.body.senha
-        }
-        
-
-        //verificar email
-        const existingEmailArray = await userModel.find({email: req.body.email});
-        const existingEmail = existingEmailArray[0];
-
-        if(existingEmail != undefined){
-            return res.status(400).send({
-                message: 'Email ja cadastrado'
-            });
+            senha: newSenha,
+            image: image
         }
 
-        //verificar cpf
-        const existingCPFArray = await userModel.find({CPF: newUser.CPF});
-        const existingCPF = existingCPFArray[0];
-
-        if(existingCPF != undefined){
-            return res.status(400).send({
-                message: 'Cpf já cadastrado'
-            });
-        }
-
-        //hash de senha
         const result = await userModel.create(newUser);
 
         return res.status(201).send({
@@ -79,8 +95,8 @@ async function createUser(req, res){
 
     }catch(error){
         return res.status(400).send({
-            error: error,
-            message: 'Ocorreu um erro na criação de usuario'
+            error: error.message,
+            message: 'Ocorreu um erro na criação de usuário'
         });
     }
 }
@@ -90,11 +106,19 @@ async function updateUser(req, res) {
 
         const update = req.body;
 
-        const updatedUser = await userModel.findByIdAndUpdate(id, update);
+
+        if(update.senha) {
+            const newSenha = await crypt(update.senha);
+            update.senha = newSenha;
+        }
+        if(req.file){
+            update.image = req.file.filename;
+        }
+        const updatedUser = await userModel.findByIdAndUpdate(id, update, {returnDocument: "after"});
 
         return res.status(200).send({
             message:"Usuário atualizado com sucesso.",
-            data: update
+            data: updatedUser
         });
 
     }catch(error){
@@ -104,26 +128,48 @@ async function updateUser(req, res) {
         });
     }
 }
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 async function deleteUser(req, res){
     try{
         const id = req.params.id;
+        const user = await userModel.findById(id);
+        if(!user){
+            return res.status(400).send({
+                message: "user não encontrado"
+            });
+        }
+        const imageName = user.image;
+        
         const userDelete = await userModel.findByIdAndDelete(id);
 
         if(userDelete){
-            return res.status(200).send({
-                message: 'Usuário removido com sucesso',
-                data: userDelete
-            });
-        }else{
-            return res.status(400).send({
-                message: "Usuário não encontrado"
-            })
+            if(imageName){
+                const imagePath = path.resolve(__dirname, '..', '..','..', 'uploads', imageName);
+                // console.log(imagePath);
+                fs.unlink(imagePath, (err)=>{
+                    if(err){
+                        return res.status(400).send({
+                            message: "erro ao excluir a imagem do servidor"
+                        });
+                    }else{
+                        return res.status(200).send({
+                            message: 'Usuário removido com sucesso',
+                            data: userDelete
+                        });
+                    }
+                })
+            } else{
+
+                return res.status(200).send({
+                    message: 'User removido com sucesso',
+                    data: userDelete
+                });
+            }
         }
-        
-       
     }catch(error){
         return res.status(400).send({
-            message: 'Ocorreu um erro para remover o usuário'
+            message: 'Ocorreu um erro para remover o user'
         })
     }
 }
@@ -135,43 +181,65 @@ async function login(req, res){
 
         const existingUserArray = await userModel.find({email});
         const existingUser = existingUserArray[0];
+
         if(existingUser == undefined){
             return res.status(400).send({
                 message: "Login ou Senha errados.",
                 status: false
             });
         }
+        console.log(existingUser);
         //hash de senha
-        if(senha === existingUser.senha){
-            console.log(req.body);
-            return res.status(200).send({
-                message: "usuario autenticado com sucesso",
-                status: true,
-                token: "hahahatoken",
-                user: existingUser
-            });
-            
+        const passwordMatch = await bcrypt.compare(senha, existingUser.senha);
+        // const passwordMatch = true;
+        if(!passwordMatch){
+            throw new Error("Login ou senha errados");
         }else{
-            return res.status(400).send({
-                message: "Login ou Senha errados.",
-                status: false
-            })
+            console.log(req.body);
+            const token = jsonwebtoken.sign(
+                {
+                    id: existingUser._id,
+                    type: "user"
+                }, 
+                process.env.SECRET_JWT, 
+                {expiresIn: '15m'}
+            );
+            const { exp } = jsonwebtoken.decode(token);
+            return res.status(200).send({
+                message: "Login realizado com sucesso",
+                status: true,
+                token: token,
+                user: existingUser,
+                expiration: exp
+            });
         }
+        
     }catch(error){
         return res.status(400).send({
-            message: "ocorreu um erro no login",
+            message: error.message,
             erro: error
         });
     }
 }
 async function logout(req, res){
     try{
+        const token = req.headers.authorization;
 
+        if(token){
+            await blackListModel.create({token});
+        }
+        console.log(token)
+
+        return res.status(200).send({
+            message: "Logout efetuado com sucesso"
+        });
     }catch(error){
-        
+        return res.status(400).send({
+            message: "Ocorreu um erro ao efetuar o logout"
+        })
     }
 }
 
 
 
-export { readAll, readOne, createUser, updateUser, deleteUser, login };
+export { readAll, readOne, createUser, updateUser, deleteUser, login, logout };
